+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* Add export excel feature for list components.
 
 ## [0.9.0-beta.0] - 2017-02-22
 ### Added